Replace navigate with navigatorFactory on RouterContext
Using a navigator factory allows the current route context to be retained by
the generated `navigate` function.  This is especially useful for navigating
from event callbacks, which are invoked long after component construction and
rendering.  At that time, the route context provider will return `undefined`,
but the generated `navigate` function will still be able to navigate relative
to the route context that was in effect when the factory (or `useNavigate`)
was called to generate it.