Update README.md (#237)
Fixed the handling of the searchParams return type and convert it from String to Number in order to increment it. Default to 0 if parameter not defined yet.
The handling of the empty page String in the setSearchParams feels cleaner than initializing the page to 0, as the parameter is mostly used after the first page.